uploaded the scratch file for initial yolo v8n
This trains on coco 128 and then it also displays some of the images from the data set(s) we will use to train yolo going forward.  This is also meant to help get someone set up with pip installs.  If you have any questions about the libraries or environments, let me know.